% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_filename.R
\name{archive_filename}
\alias{archive_filename}
\title{Return a name for a archive name, based on time-region mappings and git commit
information.}
\usage{
archive_filename(reg_id, time_id, ext = ".zip")
}
\arguments{
\item{reg_id}{a character for the region id.}

\item{time_id}{a character for the time id.}

\item{ext}{the extension of the archive file (default:".zip").}
}
\value{
A character string with the archive file name, of the form
\code{data_<reg_id>_<time_id>_<commit_id>_<commit_date><ext>}.
}
\description{
Return a name for a archive name, based on time-region mappings and git commit
information.
}
