% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_restore.R
\name{archive_restore}
\alias{archive_restore}
\title{Try to save the archive in a list of directories.}
\usage{
archive_restore(filename, dir_list, extract_dir = ".")
}
\arguments{
\item{filename}{the name of the archive to save.}

\item{dir_list}{a list of potential directories where to save the archive.}

\item{extract_dir}{the directory where to unzip the archive}
}
\value{
A character string with the path of the first archive found, which
has been unzipped into \code{extract_dir}; invisibly \code{NULL} if the
archive was not found in any of the directories of \code{dir_list}.
}
\description{
Try to save the archive in a list of directories.
}
