% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gdx.R
\name{convert_gdx}
\alias{convert_gdx}
\title{Batch convert parameters and variables from a GDX}
\usage{
convert_gdx(
  gdxfile,
  reg_id,
  time_id,
  output_directory,
  region_mappings = witchtools::region_mappings,
  time_mappings = witchtools::time_mappings,
  weights = witchtools::default_weights,
  guess_input_t = "t30",
  region_name = NULL,
  guess_region = "witch17",
  default_agg_missing = "zero",
  default_meta_param = NULL
)
}
\arguments{
\item{gdxfile}{location of the GDX file.}

\item{reg_id}{final regional aggregation.}

\item{time_id}{final time_period aggregation.}

\item{output_directory}{directory where to write the converted GDX}

\item{region_mappings}{a named list of region mapping data.table.}

\item{time_mappings}{a named list of time mapping data.table.}

\item{weights}{a named list of weights used by \code{convert_region}}

\item{guess_input_t}{input time mapping if not explicitely defined}

\item{region_name}{column name of the region, reg_id if null}

\item{guess_region}{input regional mapping if not explicitely defined}

\item{default_agg_missing}{default parameter to handle missing values in
\code{convert_region}}

\item{default_meta_param}{default meta_param}
}
\value{
A character string with the path of the input \code{gdxfile}. Called
for its side effect of writing the converted GDX in \code{output_directory},
under the same base name as \code{gdxfile}.
}
\description{
\code{convert_gdx} writes a converted GDX in the \code{output_directory}.
All parameters and variables from the input \code{gdxfile} are converted
using the \code{convert_table} function. Specific conversion options
are read in the parameter \code{meta_param} also stored in the gdxfile.
}
\examples{
\dontrun{
# Not run: needs a GDX file and the gdxtools package.
convert_gdx("input/build/data_climate.gdx", "witch17", "t30", "data_witch17")
}

}
\seealso{
\code{\link{convert_table}}, \code{\link{convert_sqlite}}.

Other conversion functions:
\code{\link[=convert_duckdb]{convert_duckdb()}},
\code{\link[=convert_region]{convert_region()}},
\code{\link[=convert_sqlite]{convert_sqlite()}},
\code{\link[=convert_table]{convert_table()}},
\code{\link[=convert_time_period]{convert_time_period()}}
}
\concept{conversion functions}
