% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_weights}
\alias{default_weights}
\title{Default weights.}
\format{
A list of data.table:
\describe{
\item{pop}{Population in 2005}
\item{gdp}{GDP in 2005}
\item{co2}{CO2 FFI in 2005}
\item{extr_coal_2000}{Coal extraction in 2000}
\item{extr_gas_2000}{Gas extraction in 2000}
\item{extr_oil_2000}{Oil extraction in 2000}
\item{tpes_2005}{TPES in 2005}
\item{prodelec_2005}{Electricity production in 2005}
\item{prodelec_hydro_2005}{Electricity production from hydroplants in 2005}
\item{extr_oil_gas_2000}{Extraction of oil and gas in 2000}
\item{forest}{Forest cover}
\item{land}{Forest cover}
\item{cst}{constant}

}
}
\source{
\url{https://tntcat.iiasa.ac.at/SspDb},
\url{https://www.pik-potsdam.de/paris-reality-check/primap-hist/},
\url{http://cait.wri.org/},
\url{https://www.iea.org/reports/world-energy-outlook-2018},
\url{https://databank.worldbank.org/source/world-development-indicators}
}
\usage{
default_weights
}
\description{
A collection of country-level Statistics to be used as weights
for regional disaggregation/aggregation. Values are given for 250 ISO3.
}
\keyword{datasets}
