% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_sets.R
\name{eu27_regions}
\alias{eu27_regions}
\title{EU27 regions for a given region mapping.}
\usage{
eu27_regions(region_mapping)
}
\arguments{
\item{region_mapping}{a data.table of regional mapping.}
}
\value{
a vector of region name.
}
\description{
\code{eu27_region} returns a vector of region representing EU27 member states. The
selection is based on the GDP of countries. If the total GDP of EU27
countries from the region is greater than 50\% of total GDP of the region,
the region is considered as belonging of Europe.
}
\examples{
eu27_regions(region_mappings[["witch17"]])
}
\seealso{
Other misc functions:
\code{\link[=brazil_regions]{brazil_regions()}},
\code{\link[=china_regions]{china_regions()}},
\code{\link[=eu28_regions]{eu28_regions()}},
\code{\link[=eu_regions]{eu_regions()}},
\code{\link[=europe_regions]{europe_regions()}},
\code{\link[=india_regions]{india_regions()}},
\code{\link[=indonesia_regions]{indonesia_regions()}},
\code{\link[=japan_regions]{japan_regions()}},
\code{\link[=korea_regions]{korea_regions()}},
\code{\link[=oceania_regions]{oceania_regions()}},
\code{\link[=oecd_regions]{oecd_regions()}},
\code{\link[=region_id]{region_id()}},
\code{\link[=southafrica_regions]{southafrica_regions()}},
\code{\link[=ssa_regions]{ssa_regions()}},
\code{\link[=usa_regions]{usa_regions()}}
}
\concept{misc functions}
