% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_scenario.R
\name{guess_scenario}
\alias{guess_scenario}
\title{Guess scenario names from a list of results gdx files.}
\usage{
guess_scenario(gdxlist, keep_ssp = FALSE, sep = "_")
}
\arguments{
\item{gdxlist}{list of results gdx files}

\item{keep_ssp}{if TRUE, keep the ssp in the name even if it the same across
scenario}

\item{sep}{separator between the directory prefix and the name, if any.}
}
\description{
Return a data.table with two columns: gdx and scenario. The scenario
is obtained from the name of the gdx without 'results_' and the extension.
If the files are located in different directory
}
