% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamc_native_regions_yml.R
\name{iamc_native_regions_yml}
\alias{iamc_native_regions_yml}
\title{Write the IAMC native regions in yaml as defined in
https://github.com/IAMconsortium/common-definitions}
\usage{
iamc_native_regions_yml(filename = NULL, model = "WITCH 5.0", n = "witch17")
}
\arguments{
\item{filename}{yaml filename to write}

\item{model}{model version}

\item{n}{Witch region definition, as defined in witchtools.}
}
\value{
Invisibly \code{NULL}. Called for its side effect of writing the
native regions yaml file to \code{filename}, or to
\code{native_regions_<model>.yml} in the working directory when
\code{filename} is \code{NULL}.
}
\description{
\code{iamc_native_regions_yml} write a yaml file with the native regions.
}
\author{
Lara Aleluia Reis
}
