% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_region.R
\name{load_region_definition}
\alias{load_region_definition}
\title{Load region definition from a regional WITCH GAMS file.}
\usage{
load_region_definition(f)
}
\arguments{
\item{f}{Regional WITCH GAMS file}
}
\description{
Load region definition from a regional WITCH GAMS file.
}
\examples{
\dontrun{
load_region_definition('input/regions/witch17.inc')
}

}
