% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_region.R
\name{load_region_mapping}
\alias{load_region_mapping}
\title{Load region mapping from a regional WITCH GAMS file.}
\usage{
load_region_mapping(f)
}
\arguments{
\item{f}{Regional WITCH GAMS file}
}
\description{
Load region mapping from a regional WITCH GAMS file.
}
\examples{
\dontrun{
load_region_mapping('input/regions/witch17.inc')
}

}
