% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_timescale.R
\name{load_timescale_mapping}
\alias{load_timescale_mapping}
\title{Load timescale mapping from a timescale CSV file.}
\usage{
load_timescale_mapping(f)
}
\arguments{
\item{f}{times WITCH GAMS file}
}
\description{
Load timescale mapping from a timescale CSV file.
}
\examples{
\dontrun{
load_timescale_mapping('input/time/t30.inc')
}

}
