% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{load_weights}
\alias{load_weights}
\title{Load weights using witch-data folder}
\usage{
load_weights(idir, region_mappings)
}
\arguments{
\item{idir}{input data directory}

\item{region_mappings}{named list of region mappings}
}
\description{
Load weights using witch-data folder
}
\examples{
\dontrun{
path <- system.file("regions", package = "witchtools")
files <- Sys.glob(file.path(path,"*.inc"))
region_mappings <- lapply(files, load_region_mapping)
names(region_mappings) <- stringr::str_sub(basename(files), 1, -5)
load_weights('../witch-data',load_region_mapping(region_mappings))
}

}
