% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_pal.R
\name{many_pals}
\alias{many_pals}
\title{Many color palettes}
\usage{
many_pals(component = NULL, variant = NULL, theme = NULL, include_names = NULL)
}
\arguments{
\item{component}{palette's component (\code{fuel}, \code{region}).}

\item{theme}{A palette theme (\code{witch-plot} (default)).}

\item{restrict_names}{only return the corresponding named values.}
}
\description{
Creates thematic palettes for WITCH results and other
}
