% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_pal.R
\name{paleiee}
\alias{paleiee}
\title{Additional Color palettes (in test)}
\usage{
paleiee(component = NULL, theme = NULL)
}
\arguments{
\item{component}{palette's component (\code{fuel}, \code{region}).}

\item{theme}{A palette theme (\code{witch-plot} (default)).}
}
\description{
Creates thematic palettes
}
