% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premise_region_mapping.R
\name{premise_region_mapping}
\alias{premise_region_mapping}
\title{PREMISE region mapping.}
\usage{
premise_region_mapping(n = "witch17", filename = NULL)
}
\arguments{
\item{n}{regional mapping ID}

\item{filename}{the name of the file to save the json}
}
\value{
Invisibly \code{NULL}. Called for its side effect of writing the
json mapping to \code{filename}, or of printing it to the console when
\code{filename} is \code{NULL}.
}
\description{
\code{premise_region_mapping} builds a json containing a regional
mapping to be used by PREMISE. It is written to \code{filename}, or
printed to the console if no filename is provided.
}
\examples{
# Write the PREMISE topology for the default region mapping
premise_region_mapping(filename = tempfile(fileext = ".json"))

}
