% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_descriptions}
\alias{region_descriptions}
\title{Default regional descriptions}
\format{
A named list of data.tables with 2 columns
\describe{
\item{mapping_name}{Mapping name (example, witch17),
contains the region names in lower case}
\item{description}{Region description}
}
}
\usage{
region_descriptions
}
\description{
A collection of regional descriptions for 250 ISO3.
}
\keyword{datasets}
