% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_mappings}
\alias{region_mappings}
\title{Default regional mapping}
\format{
A named list of data.tables with 2 columns
\describe{
\item{mapping_name}{Mapping name (example, witch17),
contains the region names in lower case}
\item{iso3}{ISO3 code in upper case}
}
}
\usage{
region_mappings
}
\description{
A collection of regional mappings for 250 ISO3.
}
\keyword{datasets}
