% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_mappings}
\alias{time_mappings}
\title{Default timescale mapping}
\format{
A named list of data.tables with 7 columns
\describe{
\item{t}{time period as a ordered numeric ID}
\item{year}{time year}
\item{refyear}{reference year for the time period}
\item{pred}{predecessor ID (t) of the time period}
\item{tperiod}{time step of the time period, used for stochastic branch,
equal to t when deterministic}
\item{begyear}{first year of the time period}
\item{endyear}{last year of the time period}
}
}
\usage{
time_mappings
}
\description{
A collection of timescale mappings.
}
\keyword{datasets}
