% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{witch_meta_param}
\alias{witch_meta_param}
\title{WITCH default meta parameter}
\format{
A data.table with 3 columns
\describe{
\item{parameter}{Name of the parameter or the variable}
\item{type}{meta-data parameter, here only nagg or nweight}
\item{value}{value of the meta-data parameter}
}
}
\usage{
witch_meta_param
}
\description{
A table containing default behavior for the WITCH parameters and variables
for regional conversion.
}
\keyword{datasets}
