% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_pal.R
\name{witch_pal}
\alias{witch_pal}
\title{WITCH palettes for WITCH results}
\usage{
witch_pal(theme = NULL, restrict = NULL)
}
\arguments{
\item{theme}{A theme palette name.}

\item{restrict}{only return the values with name in this vector.}

\item{components}{.}
}
\description{
Creates thematic palettes for WITCH results
}
\seealso{
\code{\link{witch_query}}, \code{\link{witch_scen}}.

Other WITCH helper functions: 
\code{\link{witch_query}()},
\code{\link{witch_results_files}()},
\code{\link{witch_scen_name}()}
}
\concept{WITCH helper functions}
