% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_results_files.R
\name{witch_results_files}
\alias{witch_results_files}
\title{List of WITCH results files}
\usage{
witch_results_files(
  search_path,
  restrict = "^results_",
  normalize = FALSE,
  recursive = FALSE
)
}
\arguments{
\item{search_path}{Path where to start the search of the results files (default="^results_").}

\item{restrict}{vector of filtering patterns on the file names.}

\item{normalize}{if TRUE, return canonical form of the file names.}

\item{recursive}{if TRUE, search also in the sub directories.}
}
\value{
A character vector with the unique names of the gdx files found in
\code{search_path} whose base name matches one of the \code{restrict}
patterns. Returns \code{NULL} if no file matches.
}
\description{
Returns a vector of WITCH results files. Inspired by witch-plot source code.
}
\seealso{
\link{witch_query}.

Other WITCH helper functions:
\code{\link[=witch_query]{witch_query()}},
\code{\link[=witch_scen_name]{witch_scen_name()}}
}
\concept{WITCH helper functions}
