% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_scen.R
\name{witch_scen_name}
\alias{witch_scen_name}
\title{Scenario name from the GDX filename}
\usage{
witch_scen_name(gdxfilename)
}
\arguments{
\item{gdxfilename}{path to a results gdx file.}
}
\value{
A character string with the scenario name.
}
\description{
Returns a scenario name
}
\seealso{
Other WITCH helper functions:
\code{\link[=witch_query]{witch_query()}},
\code{\link[=witch_results_files]{witch_results_files()}}
}
\concept{WITCH helper functions}
