% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witchtools-package.R
\docType{package}
\name{witchtools-package}
\alias{witchtools}
\alias{witchtools-package}
\title{witchtools: Data Management for Integrated Assessment Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools and scripts useful for integrated assessment models. It converts input datasets in time and in regions, it provides diagnostic and validation tools and it provides easy access to outputs. The package has been developed in the WITCH modelling framework but it is not limited to it.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://witchtools.witchmodel.org}
  \item Report bugs at \url{https://github.com/witch-team/witchtools/issues}
}

}
\author{
\strong{Maintainer}: Laurent Drouet \email{ldrouet@gmail.com} (\href{https://orcid.org/0000-0002-4087-7662}{ORCID})

Authors:
\itemize{
  \item Laurent Drouet \email{ldrouet@gmail.com} (\href{https://orcid.org/0000-0002-4087-7662}{ORCID})
  \item Johannes Emmerling \email{johannes.emmerling@cmcc.it} (\href{https://orcid.org/0000-0003-0916-9913}{ORCID})
}

Other contributors:
\itemize{
  \item Giacomo Marangoni \email{jackjackk@gmail.com} [contributor]
  \item Pietro Andreoni \email{pietro.andreoni@eiee.org} [contributor]
  \item Lara Aleluia Reis \email{laleluia@gmail.com} [contributor]
  \item Marco Gambarini \email{marco.gambarini@cmcc.it} [contributor]
}

}
\keyword{internal}
