% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_native_regions_yml.R
\name{write_native_regions_yml}
\alias{write_native_regions_yml}
\title{Write the IAMC native regions in yaml as defined in
https://github.com/IAMconsortium/common-definitions}
\usage{
write_native_regions_yml(
  model = "WITCH 5.0",
  regdef = "witch17",
  filename = NULL
)
}
\arguments{
\item{model}{model version}

\item{regdef}{Witch region definition. THE REGION MUST BE DEFINED IN WTCHTOOLS.}

\item{filename}{yaml filename to write}
}
\description{
\code{write_native_regions_yml} write a yaml file with the native regions.
}
\author{
Lara Aleluia Reis
}
