% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_region_mappings_yml.R
\name{write_region_mappings_yml}
\alias{write_region_mappings_yml}
\title{Write the IAMC regional mapping to common regions for WITCH yml file as in
https://github.com/IAMconsortium/common-definitions/blob/main/mappings/REMIND_3.1.yml}
\usage{
write_region_mappings_yml(
  model = "WITCH 5.0",
  regdef = "witch17",
  filename = NULL
)
}
\arguments{
\item{model}{model version}

\item{regdef}{Witch region definition. THE REGION MUST BE DEFINED IN WTCHTOOLS.}

\item{filename}{yaml filename to write}
}
\description{
\code{make_reg_mapping_yml} write a yaml file with the region mappings.
}
\author{
Lara Aleluia Reis
}
