% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_store.R
\name{archive_store}
\alias{archive_store}
\title{Try to save the archive in a list of directories.}
\usage{
archive_store(filename, dir_list)
}
\arguments{
\item{filename}{the name of the archive to save.}

\item{dir_list}{a list of potential directories where to save the archive.}
}
\description{
Try to save the archive in a list of directories.
}
