% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_region.R
\name{convert_region}
\alias{convert_region}
\title{Convert a data.table from a regional mapping to another.}
\usage{
convert_region(
  .x,
  from_reg,
  to_reg,
  agg_operator = "sum",
  agg_weight = default_weights[["gdp"]],
  agg_missing = "NA",
  regions = region_mappings,
  value_name = "value",
  region_name = "n"
)
}
\arguments{
\item{.x}{a well-formatted data.table.}

\item{from_reg}{initial regional mapping name or a data.table with
the mapping.}

\item{to_reg}{final regional mapping name  or a data.table with the mapping.}

\item{agg_operator}{aggregation operator (See Details for the list of
possible values)}

\item{agg_weight}{aggregation weight data.table (See Details for the list of
possible values)}

\item{agg_missing}{tells how to deal with missing values ("NA" or "zero")}

\item{regions}{optional list of region mappings (see Details for format)}

\item{value_name}{string column name for value.}

\item{region_name}{string column name for region.}
}
\value{
a list containing a converted data.table and information about
the coperture if available.
}
\description{
\code{convert_region} returns a list containing a data.table where values are
converted from one regional mapping to another. The conversion is done by
downscaling values at country-level (if necessary) and updscaling them
into the final regional mapping. The list also contain an
optional data.table about the value coperture for the operator \code{sumby}.
The required format of the input data is described in Details.
}
\details{
The input data.table should contain a column "value" and a column named by
the initial regional mapping \code{from_reg},
while the other columns are are considered as id columns.
The resulting data.table have a column named by the final regional
mapping \code{to_reg}.
Note that the value column name can be respecified.

The regional mapping for \code{from_reg} and \code{to_reg} can be
provided in a named list through the parameter \code{regions}.
Regional mappings are 2-columns data.table with a column named 'iso3'
(for country ISO3)
and another one named as the regional mapping (for region name).
The name in the list should also be the regional mapping name.
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{convert_DT}} for data.table,
\code{\link{convert_gdx}} for WITCH gdx files.

Other conversion functions: 
\code{\link{convert_table}()},
\code{\link{convert_time_period}()}
}
\concept{conversion functions}
