% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_sets.R
\name{eu28_regions}
\alias{eu28_regions}
\title{EU28 regions for a given region mapping.}
\usage{
eu28_regions(region_mapping)
}
\arguments{
\item{region_mapping}{a data.table of regional mapping.}
}
\value{
a vector of region name.
}
\description{
\code{eu28_region} returns a vector of region representing the former EU28 member states. The
selection is based on the GDP of countries. If the total GDP of EU28
countries from the region is greater than 50\% of total GDP of the region,
the region is considered as belonging of Europe.
}
\examples{
eu28_regions(region_mappings[["witch17"]])
}
\seealso{
Other misc functions: 
\code{\link{brazil_regions}()},
\code{\link{china_regions}()},
\code{\link{eu27_regions}()},
\code{\link{eu_regions}()},
\code{\link{europe_regions}()},
\code{\link{india_regions}()},
\code{\link{indonesia_regions}()},
\code{\link{oceania_regions}()},
\code{\link{oecd_regions}()},
\code{\link{region_id}()},
\code{\link{ssa_regions}()},
\code{\link{usa_regions}()}
}
\concept{misc functions}
