% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_sets.R
\name{india_regions}
\alias{india_regions}
\title{India regions for a given region mapping.}
\usage{
india_regions(region_mapping)
}
\arguments{
\item{region_mapping}{a data.table of regional mapping.}
}
\value{
a vector of region name.
}
\description{
\code{india_regions} returns a vector of region representing China.
}
\examples{
india_regions(region_mappings[["witch17"]])
}
\seealso{
Other misc functions: 
\code{\link{brazil_regions}()},
\code{\link{china_regions}()},
\code{\link{eu27_regions}()},
\code{\link{eu28_regions}()},
\code{\link{eu_regions}()},
\code{\link{oecd_regions}()},
\code{\link{region_id}()},
\code{\link{usa_regions}()}
}
\concept{misc functions}
