% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_libraries.R
\name{install_duckdb_safe}
\alias{install_duckdb_safe}
\title{Install duckdb package with proper handling}
\usage{
install_duckdb_safe()
}
\description{
Install duckdb package with proper handling
}
\keyword{internal}
