% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_libraries.R
\name{install_package_safe}
\alias{install_package_safe}
\title{Install a package with proper handling of non-interactive sessions}
\usage{
install_package_safe(pkg)
}
\arguments{
\item{pkg}{package name as a character string}
}
\description{
Install a package with proper handling of non-interactive sessions
}
\keyword{internal}
