% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{load_weights}
\alias{load_weights}
\title{Load weights using witch-data folder}
\usage{
load_weights(idir, region_mappings)
}
\arguments{
\item{idir}{input data directory, where data are downloaded or already present}

\item{region_mappings}{named list of region mappings}
}
\description{
Load weights using witch-data folder
}
\examples{
\dontrun{
region_mapping_files = Sys.glob(file.path(system.file("regions", package = "witchtools"),"*.inc"))
region_mappings <- lapply(region_mapping_files, load_region_mapping)
names(region_mappings) <- stringr::str_sub(basename(region_mapping_files), 1, -5)
load_weights('../witch-data',load_region_mapping(region_mappings))
}

}
