% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_sets.R
\name{oecd_regions}
\alias{oecd_regions}
\title{OECD regions for a given region mapping.}
\usage{
oecd_regions(region_mapping)
}
\arguments{
\item{region_mapping}{a data.table of regional mapping.}
}
\value{
a vector of region name.
}
\description{
\code{oecd_region} returns a vector of region representing OECD. The
selection is based on the GDP of countries. If the total GDP of OECD
countries from the region is greater than 50\% of total GDP of the region,
the region is considered as belonging of the OECD.
}
\examples{
oecd_regions(region_mappings[["witch17"]])
}
\seealso{
Other misc functions: 
\code{\link{brazil_regions}()},
\code{\link{china_regions}()},
\code{\link{eu27_regions}()},
\code{\link{eu28_regions}()},
\code{\link{eu_regions}()},
\code{\link{europe_regions}()},
\code{\link{india_regions}()},
\code{\link{region_id}()},
\code{\link{usa_regions}()}
}
\concept{misc functions}
