% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premise_region_mapping.R
\name{premise_region_mapping}
\alias{premise_region_mapping}
\title{PREMISE region mapping.}
\usage{
premise_region_mapping(n = "witch17", filename = NULL)
}
\arguments{
\item{n}{regional mapping ID}

\item{filename}{the name of the file to save the json}
}
\description{
\code{premise_region_mapping} returns a json containing a regional
mapping to be used be PREMISE. If the filename is not provided, it
will return the json as text.
}
\examples{
\dontrun{
premise_region_mapping(filename = "witch-topology.json")
}

}
