% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_id.R
\name{region_id}
\alias{region_id}
\title{Region ID of a regional mapping.}
\usage{
region_id(region_mapping)
}
\arguments{
\item{region_mapping}{a data.table of regional mapping.}
}
\value{
a character for the regional mapping ID.
}
\description{
\code{region_id} returns the region id of a region mapping.
}
\examples{
region_id(region_mappings[["witch17"]])
}
\seealso{
Other misc functions:
\code{\link[=brazil_regions]{brazil_regions()}},
\code{\link[=china_regions]{china_regions()}},
\code{\link[=eu27_regions]{eu27_regions()}},
\code{\link[=eu28_regions]{eu28_regions()}},
\code{\link[=eu_regions]{eu_regions()}},
\code{\link[=europe_regions]{europe_regions()}},
\code{\link[=india_regions]{india_regions()}},
\code{\link[=indonesia_regions]{indonesia_regions()}},
\code{\link[=japan_regions]{japan_regions()}},
\code{\link[=korea_regions]{korea_regions()}},
\code{\link[=oceania_regions]{oceania_regions()}},
\code{\link[=oecd_regions]{oecd_regions()}},
\code{\link[=southafrica_regions]{southafrica_regions()}},
\code{\link[=ssa_regions]{ssa_regions()}},
\code{\link[=usa_regions]{usa_regions()}}
}
\concept{misc functions}
