% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_id.R
\name{region_id}
\alias{region_id}
\title{Region ID of a regional mapping.}
\usage{
region_id(region_mapping)
}
\arguments{
\item{region_mapping}{a data.table of regional mapping.}
}
\value{
a character for the regional mapping ID.
}
\description{
\code{region_id} returns the region id of a region mapping.
}
\examples{
region_id(region_mappings[["witch17"]])
}
\seealso{
Other misc functions: 
\code{\link{eu27_regions}()},
\code{\link{eu28_regions}()},
\code{\link{eu_regions}()},
\code{\link{oecd_regions}()}
}
\concept{misc functions}
