% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_libraries.R
\name{require_gdxtools}
\alias{require_gdxtools}
\title{Install, if necessary, and load the gdxtools library}
\usage{
require_gdxtools(loading = TRUE)
}
\arguments{
\item{loading}{if TRUE, package is loaded}
}
\description{
Install, if necessary, and load the gdxtools library
}
\examples{
\dontrun{
require_gdxtools()
}
}
