% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_libraries.R
\name{require_package}
\alias{require_package}
\title{Install, if necessary, and load R libraries}
\usage{
require_package(pkgs, loading = TRUE)
}
\arguments{
\item{pkgs}{package names as a character vector}

\item{loading}{if TRUE, package is loaded}
}
\description{
Install, if necessary, and load R libraries
}
\examples{
\dontrun{
require_package(c("data.table","gdxtools"), loading = FALSE)
}
}
