% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_sets.R
\name{usa_regions}
\alias{usa_regions}
\title{USA regions for a given region mapping.}
\usage{
usa_regions(region_mapping)
}
\arguments{
\item{region_mapping}{a data.table of regional mapping.}
}
\value{
a vector of region name.
}
\description{
\code{usa_regions} returns a vector of region representing USA.
}
\examples{
usa_regions(region_mappings[["witch17"]])
}
\seealso{
Other misc functions:
\code{\link[=brazil_regions]{brazil_regions()}},
\code{\link[=china_regions]{china_regions()}},
\code{\link[=eu27_regions]{eu27_regions()}},
\code{\link[=eu28_regions]{eu28_regions()}},
\code{\link[=eu_regions]{eu_regions()}},
\code{\link[=europe_regions]{europe_regions()}},
\code{\link[=india_regions]{india_regions()}},
\code{\link[=indonesia_regions]{indonesia_regions()}},
\code{\link[=japan_regions]{japan_regions()}},
\code{\link[=korea_regions]{korea_regions()}},
\code{\link[=oceania_regions]{oceania_regions()}},
\code{\link[=oecd_regions]{oecd_regions()}},
\code{\link[=region_id]{region_id()}},
\code{\link[=southafrica_regions]{southafrica_regions()}},
\code{\link[=ssa_regions]{ssa_regions()}}
}
\concept{misc functions}
