% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_data.R
\name{witch_data}
\alias{witch_data}
\title{Handler for witch-data files in WITCH}
\usage{
witch_data(
  file,
  version = NULL,
  idir = getOption("witchtools.idir"),
  method = getOption("witchtools.method"),
  noCheck = getOption("witchtools.noCheck", FALSE),
  repo = getOption("witchtools.witch_data_repo"),
  remote = getOption("witchtools.witch_data_remote")
)
}
\arguments{
\item{file}{Name of the file in the with-data repository}

\item{version}{File version, only required for method piggyback (deprecated)}

\item{idir}{directory to read/download files}

\item{method}{'dvc' or 'local'.}

\item{noCheck}{don't check and download the file.}

\item{repo}{github repository name}

\item{remote}{dvc remote storage}
}
\description{
Returns the location of a file from the witch-data repository for processing,
method (\code{local}) or after checking/downloading it from the DVC remote drive
(method \code{dvc}).
}
\examples{
\dontrun{
witch_data("ssp/ssp_population.csv", version = "v0.0.1")
witch_data("ssp/ssp_population.csv", method = "witch-data")
witch_data("ssp/ssp_population.csv", noCheck = TRUE)
}
}
