% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_data.R
\name{witch_data_upload}
\alias{witch_data_upload}
\title{Upload a file as an asset in a release of a github repository.}
\usage{
witch_data_upload(
  file,
  version = NULL,
  method = getOption("witchtools.method"),
  repo = getOption("witchtools.witch_data_repo")
)
}
\arguments{
\item{file}{Name of the file in the with-data repository}

\item{version}{Release version of the file (without space)}

\item{method}{'piggyback'.}

\item{repo}{github repository name}
}
\description{
Upload a file as an asset in a release of a github repository.
}
\examples{
\dontrun{
witch_data_upload("ssp/ssp_population.csv", version = 'v0.0.1')
}
}
