% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_data_usage.R
\name{witch_data_usage}
\alias{witch_data_usage}
\title{Wrapper for the usage of a make_data file in WITCH for future use of witch_data}
\usage{
witch_data_usage(file)
}
\arguments{
\item{file}{Name of the make data file}
}
\description{
Returns the list containing the options as returned by docopt, the method
name \code{method} and the location of the input data directory. It also updates
the witchtools options, according to the call parameters.
}
\examples{
\dontrun{
opts <- witch_data_usage("make_data_validation.R")
}
}
