% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_pal.R
\name{witch_pal}
\alias{witch_pal}
\title{WITCH palettes for WITCH results}
\usage{
witch_pal(theme = NULL, include_names = NULL)
}
\arguments{
\item{name}{A palette name from the lists below.}

\item{restrict_names}{only return the values with name in this vector.}
}
\description{
Creates thematic palettes for WITCH results
}
\seealso{
\code{\link{witch_query}}, \code{\link{witch_scen}}.

Other WITCH helper functions: 
\code{\link{witch_query}()},
\code{\link{witch_scen_name}()}
}
\concept{WITCH helper functions}
