% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_query.R
\name{witch_query}
\alias{witch_query}
\title{Fast query of WITCH results files}
\usage{
witch_query(
  item,
  idx,
  resgdx,
  scen = TRUE,
  scen_fun = witch_scen_name,
  valigdx = NULL,
  histgdx = NULL,
  clean = TRUE
)
}
\arguments{
\item{item}{parameter or variable name}

\item{idx}{named list of queries}
}
\description{
Returns a data.table containing the result of the query with additional
columns on scenario
}
