% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_query.R
\name{witch_query}
\alias{witch_query}
\title{Fast query of WITCH results files}
\usage{
witch_query(
  item,
  resgdx,
  filter = list(),
  scenarios = guess_scenario(resgdx),
  add_year = "t30",
  keep_gdx = FALSE,
  keep_t = FALSE,
  year_mapping = witch_period_year,
  valigdx = NULL,
  histgdx = NULL,
  ...
)
}
\arguments{
\item{item}{parameter or variable name}

\item{resgdx}{list of WITCH results gdx}

\item{filter}{named list of filter (eg. list(e="CO2",n="brazil,usa")).
if n contains "world", then the sum of n is computed.}

\item{scenarios}{vector of scenario names in same order than resgdx}

\item{add_year}{convert t into year}

\item{keep_gdx}{keep gdx file name in the result}

\item{keep_t}{keep t in the result}

\item{year_mapping}{a mapping table to translate t into year}

\item{...}{additional parameters to send to batch_extract}

\item{scen_table}{a conversion function or a mapping table to translate gdx into scenario name}
}
\description{
Returns a formatted data.table from a list of results files.
It adds a scenario column
}
\seealso{
Other WITCH helper functions: 
\code{\link{witch_results_files}()},
\code{\link{witch_scen_name}()}
}
\concept{WITCH helper functions}
