% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_query.R
\name{witch_query}
\alias{witch_query}
\title{Fast query of WITCH results files}
\usage{
witch_query(
  item = NULL,
  resgdx = NULL,
  filter = list(),
  add_scenario = TRUE,
  scenario_mapping = witch_scen_name,
  add_year = TRUE,
  year_mapping = witch_period_year,
  valigdx = NULL,
  histgdx = NULL,
  agg_n = NULL,
  clean_columns = TRUE,
  ...
)
}
\arguments{
\item{item}{parameter or variable name}

\item{resgdx}{list of results gdx from WITCH}

\item{filter}{named list of queries}

\item{add_year}{convert t into year}

\item{year_mapping}{a mapping table to translate t into year}

\item{add_scen}{convert gdx into scenario name}

\item{scen_table}{a conversion function or a mapping table to translate gdx into scenario name}
}
\description{
Returns a data.table containing the result of the query with additional
columns on scenario
}
\seealso{
\code{\link{witch_scen}}, \code{\link{witch_pal}}.

Other WITCH helper functions: 
\code{\link{witch_pal}()},
\code{\link{witch_results_files}()},
\code{\link{witch_scen_name}()}
}
\concept{WITCH helper functions}
