% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_results_files.R
\name{witch_results_files}
\alias{witch_results_files}
\title{List of WITCH results files}
\usage{
witch_results_files(
  search_path,
  restrict = "^results_",
  normalize = FALSE,
  recursive = FALSE
)
}
\arguments{
\item{search_path}{Path where to start the search of the results files (default="^results_").}

\item{restrict}{vector of filtering patterns on the file names.}

\item{normalize}{if TRUE, return canonical form of the file names.}

\item{recursive}{if TRUE, search also in the sub directories.}
}
\description{
Returns a vector of WITCH results files. Inspired by witch-plot source code.
}
\seealso{
\code{\link{witch_query}}, \code{\link{witch_scen}}.

Other WITCH helper functions: 
\code{\link{witch_pal}()},
\code{\link{witch_query}()},
\code{\link{witch_scen_name}()}
}
\concept{WITCH helper functions}
