% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_scen.R
\name{witch_scen_name}
\alias{witch_scen_name}
\title{Scenario name from the GDX filename}
\usage{
witch_scen_name(gdxfilename)
}
\description{
Returns a scenario name
}
\seealso{
\code{\link{witch_query}}, \code{\link{witch_pal}}.

Other WITCH helper functions: 
\code{\link{witch_pal}()},
\code{\link{witch_query}()},
\code{\link{witch_results_files}()}
}
\concept{WITCH helper functions}
