% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_time_mapping.R
\name{witch_time_mapping}
\alias{witch_time_mapping}
\title{Read time mapping from a csv file.}
\usage{
witch_time_mapping(f)
}
\arguments{
\item{f}{time mapping csv file}
}
\description{
\code{witch_time_mapping} reads a time csv file and returns the
time mapping as defined in the file. The function scans the csv file.
}
\examples{
\dontrun{
witch_time_mapping("input/time/t30.csv")
}

}
