% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_translate_data.R
\name{witch_translate_data}
\alias{witch_translate_data}
\title{Translate WITCH data (run make_data files and convert data).}
\usage{
witch_translate_data(
  witch_dir = ".",
  region,
  timescale,
  idir = NULL,
  output_dir = NULL,
  regions = region_mappings,
  times = time_mappings
)
}
\arguments{
\item{witch_dir}{WITCH main directory}

\item{region}{final regional aggregation}

\item{timescale}{final timescale aggregation}

\item{idir}{input data folder (for weights and to be pass to make_data files)}

\item{output_dir}{output folder (to overidde default WITCH data folder name)}

\item{regions}{optional list of region mappings (see Details for format)}
}
\description{
\code{witch_translate_data} generates the input data for the WITCH model.
First, it will run the R \code{make_data_files} and the gams \code{make_data_files}
usually located in the input folder of WITCH. These make_data_files produce
gdx and sqlite files in the \code{build} folder. Time-series should be yearly.
Spatial data should be preferably described at ISO3 level, but the script
will handle any regional-mapping contained in \code{regions}.
}
\details{
The regional mappings should be
provided in a named list through the parameter \code{regions}.
Regional mappings are 2-columns data.table with a column named 'iso3'
(for country ISO3)
and another one named as the regional mapping (for region name).
The name in the list should also be the regional mapping name.

\code{times} is a list of time mapping between year and time period which
should be provided as a data.table with columns "year" and "t",
and refyear for interpolation and extrapolation. \code{times} should contain
\code{timescale}
}
\examples{
\dontrun{
witch_translate_data(region = "r5", timescale = 't30')
}
}
