% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/witch_translate_data.R
\name{witch_translate_data}
\alias{witch_translate_data}
\title{Translate WITCH data (run make_data files and convert data).}
\usage{
witch_translate_data(
  witch_dir = ".",
  region,
  timescale,
  idir = NULL,
  output_dir = NULL,
  regions = region_mappings,
  times = time_mappings
)
}
\arguments{
\item{witch_dir}{WITCH main directory}

\item{region}{final regional aggregation}

\item{timescale}{final timescale aggregation}

\item{idir}{input data folder (for weights and to be pass to make_data files)}

\item{output_dir}{output folder (to overidde default WITCH data folder name)}
}
\description{
Translate WITCH data (run make_data files and convert data).
}
\examples{
\dontrun{
witch_translate_data(region = "r5")
witch_translate_data(region = "r5", output_dir = 'data_r5')
}
}
