% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_native_regions_yml.R
\name{write_native_regions_yml}
\alias{write_native_regions_yml}
\title{Write the IAMC native regions in yaml as defined in
https://github.com/IAMconsortium/common-definitions}
\usage{
write_native_regions_yml(filename = NULL, model = "WITCH 5.0", n = "witch17")
}
\arguments{
\item{filename}{yaml filename to write}

\item{model}{model version}

\item{n}{Witch region definition, as defined in witchtools.}
}
\description{
\code{write_native_regions_yml} write a yaml file with the native regions.
}
\author{
Lara Aleluia Reis
}
