% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_region_mappings_yml.R
\name{write_region_mappings_yml}
\alias{write_region_mappings_yml}
\title{Write the IAMC regional mapping to common regions for WITCH yml file as in
https://github.com/IAMconsortium/common-definitions}
\usage{
write_region_mappings_yml(
  filename = NULL,
  model = "WITCH 5.0",
  n = "witch17",
  comm_regs = c("world", "r5", "r9", "r10")
)
}
\arguments{
\item{filename}{yaml filename to write}

\item{model}{model version}

\item{n}{Witch region definition. THE REGION MUST BE DEFINED IN WTCHTOOLS.}

\item{comm_regs}{Additional mappings, which region aggregate, to add.}
}
\description{
\code{write_region_mappings_yml} write a yaml file with the region mappings.
}
\author{
Lara Aleluia Reis
}
